+++ saltedge-ios-swift/Classes/SEWebView/SEWebView.swift
@@ -50,6 +50,9 @@ open class SEWebView: WKWebView {
         super.init(frame: frame, configuration: WKWebViewConfiguration())
         navigationDelegate = self
         uiDelegate = self
+        #if os(iOS)
+            tintColor = .black
+        #endif
         allowsBackForwardNavigationGestures = true
     }
     
@@ -57,6 +60,9 @@ open class SEWebView: WKWebView {
         super.init(coder: aDecoder)
         navigationDelegate = self
         uiDelegate = self
+        #if os(iOS)
+            tintColor = .black
+        #endif
         allowsBackForwardNavigationGestures = true
     }